requirements.txt
django==4.2.4
djangorestframework
django-ckeditor